package/requirements.txt
# CRUD reqs
attrs >= 18.1.0
docker
python-dateutil
redis
pyyaml
jinja2
bs4
lxml
requests
click

# Possible error
# gistsig@git+https://github.com/derekmerck/gistsig

# DIANA reqs
numpy
pydicom
pyyaml
pillow
python-dateutil
cryptography
watchdog
